Split iOS keyboard helper code
